Fix CMake module base filename
This patch renames `example.cmake` to `example-config.cmake`. This
fixes a bug in which `find_package(example)` fails to locate the
CMake module shipped by this project.

See the upstream CMake documentation for `find_package` for details
about which file basenames are accepted when searching for a
CMake modules via `find_package`, but names like `example.cmake`
are not on the search path and will not be discovered properly.

Using `example-config.cmake` resolves this issue.